refactor: modularize survey flow
