+++ src/components/EstresScreen.tsx
@@ -1,0 +1,16 @@
+import BloquesDePreguntas from "./BloquesDePreguntas";
+import { bloqueEstres, preguntasEstres } from "../data/preguntas";
+
+interface Props {
+  onFinish: (respuestas: string[]) => void;
+}
+
+export default function EstresScreen({ onFinish }: Props) {
+  return (
+    <BloquesDePreguntas
+      bloques={bloqueEstres}
+      preguntas={preguntasEstres}
+      onFinish={onFinish}
+    />
+  );
+}
