+++ src/components/SelectorScreen.tsx
@@ -1,0 +1,9 @@
+import FormSelector from "./FormSelector";
+
+interface Props {
+  onSelect: (form: "A" | "B") => void;
+}
+
+export default function SelectorScreen({ onSelect }: Props) {
+  return <FormSelector onSelect={onSelect} />;
+}
